+++ contracts/mixnet/src/mixnet_params/mod.rs
@@ -1,3 +1,4 @@
 pub mod queries;
 pub mod state;
+pub mod storage;
 pub mod transactions;
