+++ contracts/mixnet/src/mixnodes/layer_queries.rs
@@ -1,7 +1,7 @@
-use crate::storage;
+use crate::mixnet_params::storage as mixnet_params_storage;
 use cosmwasm_std::Deps;
 use mixnet_contract::LayerDistribution;
 
 pub(crate) fn query_layer_distribution(deps: Deps) -> LayerDistribution {
-    storage::read_layer_distribution(deps.storage)
+    mixnet_params_storage::read_layer_distribution(deps.storage)
 }
